feat(dash): Make all comp files entry pts, arrow funcs
-Removes event stream from mock api
