+++ web/client/src/App.scss
@@ -13,6 +13,8 @@ $small: 672px;
 @import './components/Shell/Shell.scss';
 @import './components/EventsMap/EventsMap.scss';
 @import './components/EarthquakeList/EarthquakeList.scss';
+@import './components/SaveHeader/SaveHeader.scss';
+@import './components/Field/Field.scss';
 
 body {
   background-color: #161616;
