+++ web/client/src/components/Field/index.js
@@ -1,2 +1,12 @@
-import Field from './Field.js'
+import React from 'react'
+
+const Field = ({ title, value }) => {
+  return (
+    <div className="field">
+      <p className="field_title">{title}</p>
+      <span>{value}</span>
+    </div>
+  )
+}
+
 export default Field
