--- web/client/src/components/SensorTable/index.js
@@ -1,2 +1,0 @@
-import SensorTable from './SensorTable';
-export default SensorTable;
