+++ web/client/src/components/Tile/index.js
@@ -1,7 +1,7 @@
-import React from 'react';
-import { Tile } from 'carbon-components-react';
+import React from 'react'
+import { Tile } from 'carbon-components-react'
 
-export default function index({ title, main, small }) {
+const _Tile = ({ title, main, small }) => {
   return (
     <Tile>
       <h2 className="tile_heading">{title}</h2>
@@ -10,7 +10,8 @@ export default function index({ title, main, small }) {
         <h3>{main}</h3>
         <small>{small}</small>
       </div>
-
     </Tile>
-  );
+  )
 }
+
+export default _Tile
