+++ web/client/src/components/_SensorTable/SensorTable.js
