+++ web/client/src/components/_SensorTable/index.js
@@ -1,0 +1,2 @@
+import SensorTable from './SensorTable'
+export default SensorTable
